+++ src/components/popover/popover.test.jsx
@@ -1,9 +1,22 @@
-import React, { useRef, Fragment } from 'react';
-import { fireEvent, render, screen } from '@testing-library/react';
+import React, {
+  createRef,
+  useRef,
+  Fragment,
+} from 'react';
+import {
+  act,
+  fireEvent,
+  render,
+  screen,
+} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import Popover from './popover.jsx';
 
 describe('Popover', () => {
+  const requiredProps = {
+    title: 'Test title',
+  };
+
   function PopoverWithToggle(popoverProps) {
     const popoverRef = useRef();
 
@@ -19,14 +32,8 @@ describe('Popover', () => {
 
   it('has defaults', () => {
     const popoverRef = React.createRef();
-
-    render(<Popover title="Defaults" />);
+    render(<Popover ref={popoverRef} title="Defaults" />);
     expect(document.body).toMatchSnapshot();
-
-    render(<Popover title="Defaults" startOpen={true} />);
-    expect(document.body).toMatchSnapshot();
-
-    render(<Popover ref={popoverRef} title="Defaults with ref" />);
     expect(popoverRef.current).toMatchSnapshot();
   });
 
@@ -106,7 +113,26 @@ describe('Popover', () => {
     expect(screen.getByText('Another title').tagName).toEqual('H1');
   });
 
-  describe('flush api', () => {
+  describe('autoflush API', () => {
+    // We don't know how to test this well. Manually QA for now.
+  });
+
+  describe('children API', () => {
+    it('is set', () => {
+      const ref = createRef();
+      render((
+        <Popover {...requiredProps} ref={ref}>
+          Unique popover content
+        </Popover>
+      ));
+      act(() => {
+        ref.current.open = true;
+      });
+      expect(screen.getByText('Unique popover content')).toBeInTheDocument();
+    });
+  });
+
+  describe('flush API', () => {
     it('can flush left', () => {
       render(<PopoverWithToggle title="Another title" flush="left"><span>More text</span></PopoverWithToggle>);
       userEvent.click(screen.getByText('Open Popover'));
@@ -119,6 +145,84 @@ describe('Popover', () => {
       userEvent.click(screen.getByText('Open Popover'));
       expect(screen.getByRole('dialog')).toHaveStyle({ right: 0 });
       expect(screen.getByRole('dialog')).not.toHaveStyle({ left: 0 });
+    });
+  });
+
+  describe('onClose API', () => {
+    it('is a function', () => {
+      const ref = createRef();
+      const onClose = jest.fn();
+      render(<Popover {...requiredProps} ref={ref} onClose={onClose} />);
+      act(() => {
+        ref.current.open = true;
+      });
+      expect(onClose).not.toHaveBeenCalled();
+      act(() => {
+        ref.current.open = false;
+      });
+      expect(onClose).toHaveBeenCalled();
+    });
+  });
+
+  describe('shouldClose API', () => {
+    it('allows closing via clicks', () => {
+      const ref = createRef();
+      const onClose = jest.fn();
+      const shouldClose = () => true;
+      render(<Popover {...requiredProps} ref={ref} onClose={onClose} shouldClose={shouldClose} />);
+      act(() => {
+        ref.current.open = true;
+      });
+      userEvent.click(document.body);
+      expect(onClose).toHaveBeenCalled();
+    });
+
+    it('stops closing via clicks', () => {
+      const ref = createRef();
+      const onClose = jest.fn();
+      const shouldClose = () => false;
+      render(<Popover {...requiredProps} ref={ref} onClose={onClose} shouldClose={shouldClose} />);
+      act(() => {
+        ref.current.open = true;
+      });
+      userEvent.click(document.body);
+      expect(onClose).not.toHaveBeenCalled();
+    });
+
+    it('allows closing via blur', () => {
+      const ref = createRef();
+      const onClose = jest.fn();
+      const shouldClose = () => true;
+      render(<Popover {...requiredProps} ref={ref} onClose={onClose} shouldClose={shouldClose} />);
+      act(() => {
+        ref.current.open = true;
+      });
+      userEvent.tab({ shift: true });
+      expect(onClose).toHaveBeenCalled();
+    });
+
+    it('stops closing via blur', () => {
+      const ref = createRef();
+      const onClose = jest.fn();
+      const shouldClose = () => false;
+      render(<Popover {...requiredProps} ref={ref} onClose={onClose} shouldClose={shouldClose} />);
+      act(() => {
+        ref.current.open = true;
+      });
+      userEvent.tab({ shift: true });
+      expect(onClose).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('title API', () => {
+    it('is set', () => {
+      const ref = createRef();
+      render(<Popover {...requiredProps} ref={ref} title="Unique title" />);
+      expect(screen.queryByText('Unique title')).not.toBeInTheDocument();
+      act(() => {
+        ref.current.open = true;
+      });
+      expect(screen.getByText('Unique title')).toBeInTheDocument();
     });
   });
 });