+++ CHANGELOG.md
@@ -11,6 +11,10 @@
   - feat(Form): add `className` API
 </details>
 
+## v0.71.0
+- feat(Popover): implement `shouldClose` API
+- fix(Checkbox): do not allow state change when `readOnly`
+
 ## v0.70.0
 - feat(Form): add `className` API